testing/web-platform/tests/css/css-text/i18n/css3-text-line-break-baspglwj-039.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>CSS3 Text, linebreaks: U+205E VERTICAL FOUR DOTS</title>
<link rel='author' title='Richard Ishida' href='mailto:ishida@w3.org'>
<link rel='help' href='https://drafts.csswg.org/css-text-3/#line-breaking'>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta name='flags' content='should dom'>
<meta name="assert" content="[Exploratory] The browser will break a line of text after any Unicode character with the BA historic word separator property.">
<style type='text/css'>
#breakable { font-family: monospace; font-size: 25px; width: 50ch; line-height: 30px; }
</style>
</head>
<body>



<div class="test">
			<div id="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x205E;bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div>
			</div>


<!--Notes:

Box height: <span id='debugresult'></span><script>document.getElementById('debugresult').innerHTML = document.getElementById('breakable').offsetHeight;</script>

-->
<script>
test(function() {
assert_true(document.getElementById('breakable').offsetHeight > 35);
}, " ");
</script>

<div id='log'></div>

</body>
</html>
